+++ apps/app/components/profile/overview/index.ts
@@ -1,3 +1,4 @@
+export * from "./activity";
 export * from "./priority-distribution";
 export * from "./state-distribution";
 export * from "./stats";
